ai-service/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
openai>=1.12.0
anthropic>=0.18.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
xgboost>=2.0.0
lightgbm>=4.0.0
sentence-transformers>=2.3.0
faiss-cpu>=1.7.4
httpx>=0.26.0
python-multipart>=0.0.9


